解决数据库删除表崩溃的问题；提供接口创建搜索控件；代码调整与优化

+++ lib_launcher/src/main/java/com/android/launcher3/graphics/ColorExtractor.java
@@ -37,7 +37,7 @@ public class ColorExtractor {
     public static int findDominantColorByHue(Bitmap bitmap, int samples) {
         final int height = bitmap.getHeight();
         final int width = bitmap.getWidth();
-        int sampleStride = (int) Math.sqrt((height * width) / samples);
+        int sampleStride = (int) Math.sqrt(((float) height * width) / samples);
         if (sampleStride < 1) {
             sampleStride = 1;
         }
